+++ CHANGELOG.md
@@ -3,7 +3,10 @@ hiqdev/hipanel-module-domain changelog
 
 ## Under development
 
-- moved to src, hideved
+- Added domain buy redirect
+- Fixed: many for first release
+- Changed: moved to src, hideved
+- Added basics
 - inited
 
 ## Development started 2015-04-20